checks inventory and sees if there are greater than zero glasses. Renders it on screen
